add YAML::XS::IndentlessMap for YAML.pm backcompat
Change the default indent behavior of map elements,
advance the indent the children of the map
according to the spec: 6.1 Indentation Spaces:

  "Each node must be indented further than its parent node.  All sibling nodes
  must use the exact same indentation level.  However the content of each
  sibling node may be further indented independently."
  http://yaml.org/spec/1.2/spec.html#id2777534

and make it readable by YAML,
and behave the same as all other YAML writers.

Fallback to the old YAML::XS map indent by setting the global
$YAML::XS::IndentlessMap to a true value.

Fixes RT #41463 and starts making it usable as default YAML library for core.